Update to client.js to resolve bug in watcher
Folders were not being added to chokidar (file/folder watcher) and thus nothing was being updated remotely if it was edited in a watched folder by another program. By doing the simple tweak to line 53, I tested that the folders get picked up by chokidar and the changed are sent to remote server.